test : 그룹 생성 도메인계층 테스트 코드 작성
