+++ src/test/java/com/example/surveyapi/domain/project/domain/manager/ManagerTest.java
@@ -1,0 +1,150 @@
+package com.example.surveyapi.domain.project.domain.manager;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.time.LocalDateTime;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.test.util.ReflectionTestUtils;
+
+import com.example.surveyapi.domain.project.domain.manager.entity.Manager;
+import com.example.surveyapi.domain.project.domain.manager.enums.ManagerRole;
+import com.example.surveyapi.domain.project.domain.project.entity.Project;
+import com.example.surveyapi.global.enums.CustomErrorCode;
+import com.example.surveyapi.global.exception.CustomException;
+
+public class ManagerTest {
+	@Test
+	void 매니저_추가_정상() {
+		// given
+		Project project = createProject();
+
+		// when
+		project.addManager(1L, 2L);
+
+		// then
+		assertEquals(2, project.getManagers().size());
+	}
+
+	@Test
+	void 매니저_추가_READ_권한으로_시도_실패() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.addManager(2L, 3L);
+		});
+		assertEquals(CustomErrorCode.ACCESS_DENIED, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_중복_추가_실패() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.addManager(1L, 2L);
+		});
+		assertEquals(CustomErrorCode.ALREADY_REGISTERED_MANAGER, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_권한_변경_정상() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+
+		// when
+		project.updateManagerRole(1L, 2L, ManagerRole.WRITE);
+
+		// then
+		Manager manager = project.findManagerByUserId(2L);
+		assertEquals(ManagerRole.WRITE, manager.getRole());
+	}
+
+	@Test
+	void 매니저_권한_변경_소유자가_아닌_사용자_시도_실패() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.updateManagerRole(2L, 1L, ManagerRole.WRITE);
+		});
+		assertEquals(CustomErrorCode.ACCESS_DENIED, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_권한_변경_본인_OWNER_권한_변경_시도_실패() {
+		// given
+		Project project = createProject();
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.updateManagerRole(1L, 1L, ManagerRole.WRITE);
+		});
+		assertEquals(CustomErrorCode.CANNOT_CHANGE_OWNER_ROLE, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_권한_변경_OWNER로_변경_시도_실패() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.updateManagerRole(1L, 2L, ManagerRole.OWNER);
+		});
+		assertEquals(CustomErrorCode.CANNOT_CHANGE_OWNER_ROLE, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_삭제_정상() {
+		// given
+		Project project = createProject();
+		project.addManager(1L, 2L);
+		Manager targetManager = project.findManagerByUserId(2L);
+		ReflectionTestUtils.setField(targetManager, "id", 2L);
+
+		// when
+		project.deleteManager(1L, 2L);
+
+		// then
+		assertTrue(targetManager.getIsDeleted());
+	}
+
+	@Test
+	void 존재하지_않는_매니저_ID로_삭제_실패() {
+		// given
+		Project project = createProject();
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.deleteManager(1L, 999L);
+		});
+		assertEquals(CustomErrorCode.NOT_FOUND_MANAGER, exception.getErrorCode());
+	}
+
+	@Test
+	void 매니저_삭제_본인_소유자_삭제_시도_실패() {
+		// given
+		Project project = createProject();
+		Manager ownerManager = project.findManagerByUserId(1L);
+
+		// when & then
+		CustomException exception = assertThrows(CustomException.class, () -> {
+			project.deleteManager(1L, ownerManager.getId());
+		});
+		assertEquals(CustomErrorCode.CANNOT_DELETE_SELF_OWNER, exception.getErrorCode());
+	}
+
+	private Project createProject() {
+		return Project.create("테스트", "설명", 1L, 50, LocalDateTime.now(), LocalDateTime.now().plusDays(5));
+	}
+}
